requirements.txt
discord.py
requests
yfinance
